PyJones/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    {name = "KazuhiraBenedictMiller",email = "aaserravalle.py@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
